Updates to not build the node app in the container
Building the node app in the container consumes too much resources, which is costly to do on the production environment. Instead, we publish compiled releases and only download them from the container.

To get the latest release we clone the repository and use `git describe`. This is very inefficient and we shouldn't need the whole repository to do that. But there isactually no way to properly do a remote `git describe`.